Typecheck DINO loss
